improve DB

+++ src/Entity/TournamentScore.php
@@ -79,12 +79,12 @@ class TournamentScore
         return $this;
     }
 
-    public function getScore(): ?float
+    public function getScore(): ?int
     {
         return $this->score;
     }
 
-    public function setScore(float $score): self
+    public function setScore(int $score): self
     {
         $this->score = $score;
